allow configure of system props
